Modified distribution so that TRANSVCUTOFF is now active for GUNGAUSSFLATTOPTH and ASTRAFLATTOPTH.
If TRANSVCUTOFF is negative or zero, these distribution work the same as before. (The default value for TRANSVCUTOFF is set to -3.0 and the other routine that reads it takes the absolute value.) If it is greater than zero, then sigx_m defines the Gaussian sigmas in x and y and particles greater than the radius defined by TRASNVCUTOFF are excluded. We made this change because SLAC is reporting lowered emittances using Gaussian cut transverse profiles comparted to the usual uniform profiles.