Rewrote LossDataSink compleatly.
A loss LossDataSink is initiated with the constructor, then particles are added (and locally stored).
In the finalize method all data is written to an ASCII or H5 file, depending on the NEW option ASCIIDUMP.

Testing of this is ongoing, there are maybe some gliches, watch out. 

Regressiontest will follow!

@Yianjun: please test the CYCLOTRON related PROBES and STRIPPER element!