New: Read field map information from header of the CARBONCYCL type, hence for the matched distribution the CARBONCYCL type has to be used. For example the PSI Ring header would look like:
1800.0
20.0
8
-4.0
1440
141
......

Important: the third line describes the number of sectors AND is not 
specified in the original definition of the CARBONCYCL type. THIS needs to be 
fixed as soon as the matched distribution is working properly.

At the moment the correct sigma matrix will be produced only with the ring590_bfld.dat:

 3.859  -0.6331 0     0	     0.4538 1.324

-0.6331  0.636  0     0      0.7276 -0.2605

 0	 0      6.256 1.197  0	     0

 0	 0      1.197 0.3845 0	     0

 0.4538  0.7276 0     0	     2.396   0.172

 1.324  -0.2605 0     0	     0.172   0.8693

Check at svn+ssh://savannah02.psi.ch/repos/opal/tests/trunk/RegressionTests/RingCyclotronMatched